apps/writer/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.9.0
jinja2==3.1.3
python-multipart==0.0.7
httpx==0.26.0
aiohttp==3.9.1
